refactor: 맵 쿼리 서비스 분리
