+++ src/main/java/umc/cockple/demo/domain/exercise/service/ExerciseQueryService.java
@@ -268,52 +268,6 @@ public class ExerciseQueryService {
         return exerciseConverter.toMyExerciseListResponse(exerciseSlice, participantCountMap, bookmarkStatus, isCompletedMap);
     }
 
-    public ExerciseBuildingDetailDTO.Response getBuildingExerciseDetails(
-            String buildingName, String streetAddr, LocalDate date, Long memberId) {
-
-        log.info("건물 운동 상세 조회 시작 - 건물: {}, 주소: {}, 날짜: {}", buildingName, streetAddr, date);
-
-        Member member = memberLookupService.findByIdOrThrow(memberId);
-        List<Exercise> exercises = exerciseReader.findByBuildingAndDate(buildingName, streetAddr, date);
-
-        if (exercises.isEmpty()) {
-            log.info("건물에 운동이 존재하지 않습니다. - 건물: {}, 주소: {}, 날짜: {}", buildingName, streetAddr, date);
-            return exerciseConverter.toEmptyBuildingDetailResponse(buildingName, date);
-        }
-
-        List<Long> exerciseIds = getExerciseIds(exercises);
-        Map<Long, Boolean> bookmarkStatus = exerciseBookmarkReader.getBookmarkStatus(memberId, exerciseIds);
-
-        log.info("건물 운동 상세 조회 종료 - 건물: {}, 주소: {}, 날짜: {}, 결과: {}", buildingName, streetAddr, date, exerciseIds.size());
-
-        return exerciseConverter.toBuildingDetailResponse(exercises, buildingName, bookmarkStatus, date);
-    }
-
-    public ExerciseMapBuildingsDTO.Response getExerciseMapCalendarSummary(
-            ExerciseMapBuildingsDTO.Query query, Long memberId) {
-
-        log.info("월간 운동 캘린더 요약 조회 시작 - 날짜: {}, 중심: ({}, {}), 반경: {}km",
-                query.date(), query.latitude(), query.longitude(), query.radiusKm());
-
-        Member member = memberLookupService.findWithAddressesOrThrow(memberId);
-        MemberAddr mainAddr = memberLookupService.findMainAddressOrThrow(member);
-        ExerciseMapBuildingsDTO.Query searchQuery =
-                query.withFallbackLocation(mainAddr.getLatitude(), mainAddr.getLongitude());
-
-        DateRange dateRange = DateRange.calculateMonthlyStartAndEnd(query.date());
-
-        List<Exercise> exercises = exerciseReader.findByMonthAndRadius(dateRange.start(), dateRange.end(), searchQuery);
-
-        Map<LocalDate, List<ExerciseMapBuildingsDTO.BuildingInfo>> dailyBuildings =
-                groupExercisesByDateAndBuilding(exercises);
-
-        log.info("월간 운동 캘린더 요약 조회 완료 - 조회된 운동 수: {}", exercises.size());
-
-        return exerciseConverter.toMapCalendarSummaryResponse(
-                dateRange.start().getYear(), dateRange.start().getMonthValue(),
-                searchQuery.latitude(), searchQuery.longitude(), searchQuery.radiusKm(), dailyBuildings);
-    }
-
     public ExerciseRecommendationCalendarDTO.Response getRecommendedExerciseCalendar(
             Long memberId,
             LocalDate startDate,
@@ -550,30 +504,6 @@ public class ExerciseQueryService {
                 ));
     }
 
-    private Map<LocalDate, List<ExerciseMapBuildingsDTO.BuildingInfo>> groupExercisesByDateAndBuilding(List<Exercise> exercises) {
-        Map<LocalDate, List<Exercise>> exercisesByDate = exercises.stream()
-                .collect(Collectors.groupingBy(Exercise::getDate));
-
-        return exercisesByDate.entrySet().stream()
-                .collect(Collectors.toMap(
-                        Map.Entry::getKey,
-                        entry -> createBuildingSummariesForDate(entry.getValue()),
-                        (existing, replacement) -> existing,
-                        TreeMap::new
-                ));
-    }
-
-    private List<ExerciseMapBuildingsDTO.BuildingInfo> createBuildingSummariesForDate(List<Exercise> dayExercises) {
-        Map<BuildingKey, List<Exercise>> exercisesByBuilding = dayExercises.stream()
-                .collect(Collectors.groupingBy(this::createBuildingKey));
-
-        return exercisesByBuilding.keySet().stream()
-                .map(entry -> exerciseConverter.toBuildingSummary(
-                        entry.name(), entry.address(), entry.latitude(), entry.longitude())
-                )
-                .toList();
-    }
-
     // ========== 세부 비즈니스 메서드 ==========
 
     private List<ParticipantInfo> buildMemberParticipantInfos(List<MemberExercise> memberExercises, Party party) {
@@ -722,17 +652,6 @@ public class ExerciseQueryService {
         };
     }
 
-    private BuildingKey createBuildingKey(Exercise exercise) {
-        var addr = exercise.getExerciseAddr();
-
-        return new BuildingKey(
-                addr.getBuildingName(),
-                addr.getStreetAddr(),
-                addr.getLatitude().doubleValue(),
-                addr.getLongitude().doubleValue()
-        );
-    }
-
     private record ParticipantGroups(
             List<ExerciseDetailDTO.ParticipantInfo> participants,
             List<ExerciseDetailDTO.ParticipantInfo> waiting
@@ -740,16 +659,6 @@ public class ExerciseQueryService {
     }
 
     private record DateRange(LocalDate start, LocalDate end) {
-        private static DateRange calculateMonthlyStartAndEnd(LocalDate date) {
-            LocalDate targetDate = (date != null) ? date : LocalDate.now();
-
-            LocalDate start = targetDate.withDayOfMonth(1);
-            int lastDay = targetDate.lengthOfMonth();
-            LocalDate end = targetDate.withDayOfMonth(lastDay);
-
-            return new DateRange(start, end);
-        }
-
         private static DateRange calculateDateRange(LocalDate startDate, LocalDate endDate) {
             if (startDate != null && endDate != null) {
                 return new DateRange(startDate, endDate);
@@ -767,25 +676,4 @@ public class ExerciseQueryService {
     private record ExerciseWithDistance(Exercise exercise, double distance) {
     }
 
-    private record BuildingKey(
-            String name,
-            String address,
-            Double latitude,
-            Double longitude
-    ) {
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj) return true;
-            if (obj == null || getClass() != obj.getClass()) return false;
-
-            BuildingKey that = (BuildingKey) obj;
-            return Objects.equals(name, that.name) &&
-                    Objects.equals(address, that.address);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(name, address);
-        }
-    }
 }
